Changed names.

+++ include/custom.h
@@ -1,7 +1,7 @@
 #ifndef CUSTOM_H
 #define CUSTOM_H
 
-void max_index(int *arr, size_t size, int *max, size_t *index);
-size_t *dec_to_byte(size_t value, size_t *byte);
+void max_index(int *arr, size_t size, int *max, int *index);
+int *dec_to_byte(int value, int *byte);
 
 #endif
